+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added color to Gantry messages from the entrypoint script.
+- Added `--log-level` global option.
 
 ## [v2.0.0](https://github.com/allenai/beaker-gantry/releases/tag/v2.0.0) - 2025-05-01